+++ processor/src/main/java/router/fu/processor/ProcessorUtil.java
@@ -5,15 +5,18 @@ import com.squareup.javapoet.MethodSpec;
 import com.squareup.javapoet.ParameterSpec;
 import com.squareup.javapoet.TypeSpec;
 import com.squareup.javapoet.TypeVariableName;
-import java.lang.annotation.Annotation;
 import java.lang.annotation.Documented;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import javax.lang.model.AnnotatedConstruct;
 import javax.lang.model.element.AnnotationMirror;
+import javax.lang.model.element.AnnotationValue;
 import javax.lang.model.element.Element;
 import javax.lang.model.element.ElementKind;
 import javax.lang.model.element.ExecutableElement;
@@ -24,6 +27,7 @@ import javax.lang.model.type.DeclaredType;
 import javax.lang.model.type.ExecutableType;
 import javax.lang.model.type.TypeKind;
 import javax.lang.model.type.TypeMirror;
+import javax.lang.model.util.Elements;
 import javax.lang.model.util.Types;
 
 final class ProcessorUtil
@@ -32,24 +36,27 @@ final class ProcessorUtil
   {
   }
 
-  static void mustNotThrowAnyExceptions( @Nonnull final Class<? extends Annotation> type,
+  static void mustNotThrowAnyExceptions( @Nonnull final String annotationName,
                                          @Nonnull final ExecutableElement method )
     throws RouterProcessorException
   {
     if ( !method.getThrownTypes().isEmpty() )
     {
-      throw new RouterProcessorException( "@" + type.getSimpleName() + " target must not throw any exceptions",
+      throw new RouterProcessorException( "@" +
+                                          toSimpleName( annotationName ) +
+                                          " target must not throw any exceptions",
                                           method );
     }
   }
 
-  static void mustNotHaveAnyParameters( @Nonnull final Class<? extends Annotation> type,
+  static void mustNotHaveAnyParameters( @Nonnull final String annotationName,
                                         @Nonnull final ExecutableElement method )
     throws RouterProcessorException
   {
     if ( !method.getParameters().isEmpty() )
     {
-      throw new RouterProcessorException( "@" + type.getSimpleName() + " target must not have any parameters", method );
+      throw new RouterProcessorException( "@" + toSimpleName( annotationName ) + " target must not have any parameters",
+                                          method );
     }
   }
 
@@ -57,53 +64,54 @@ final class ProcessorUtil
    * Verifies that the method is not final, static, abstract or private.
    * The intent is to verify that it can be overridden in sub-class in same package.
    */
-  static void mustBeOverridable( @Nonnull final Class<? extends Annotation> type,
+  static void mustBeOverridable( @Nonnull final String annotationName,
                                  @Nonnull final ExecutableElement method )
     throws RouterProcessorException
   {
-    mustNotBeFinal( type, method );
-    mustBeSubclassCallable( type, method );
+    mustNotBeFinal( annotationName, method );
+    mustBeSubclassCallable( annotationName, method );
   }
 
   /**
    * Verifies that the method is not static, abstract or private.
    * The intent is to verify that it can be instance called by sub-class in same package.
    */
-  private static void mustBeSubclassCallable( @Nonnull final Class<? extends Annotation> type,
+  private static void mustBeSubclassCallable( @Nonnull final String annotationName,
                                               @Nonnull final ExecutableElement method )
     throws RouterProcessorException
   {
-    mustNotBeStatic( type, method );
-    mustNotBePrivate( type, method );
+    mustNotBeStatic( annotationName, method );
+    mustNotBePrivate( annotationName, method );
   }
 
-  private static void mustNotBeFinal( @Nonnull final Class<? extends Annotation> type,
+  private static void mustNotBeFinal( @Nonnull final String annotationName,
                                       @Nonnull final ExecutableElement method )
     throws RouterProcessorException
   {
     if ( method.getModifiers().contains( Modifier.FINAL ) )
     {
-      throw new RouterProcessorException( "@" + type.getSimpleName() + " target must not be final", method );
+      throw new RouterProcessorException( "@" + toSimpleName( annotationName ) + " target must not be final", method );
     }
   }
 
-  private static void mustNotBeStatic( @Nonnull final Class<? extends Annotation> type,
+  private static void mustNotBeStatic( @Nonnull final String annotationName,
                                        @Nonnull final ExecutableElement method )
     throws RouterProcessorException
   {
     if ( method.getModifiers().contains( Modifier.STATIC ) )
     {
-      throw new RouterProcessorException( "@" + type.getSimpleName() + " target must not be static", method );
+      throw new RouterProcessorException( "@" + toSimpleName( annotationName ) + " target must not be static", method );
     }
   }
 
-  private static void mustNotBePrivate( @Nonnull final Class<? extends Annotation> type,
+  private static void mustNotBePrivate( @Nonnull final String annotationName,
                                         @Nonnull final ExecutableElement method )
     throws RouterProcessorException
   {
     if ( method.getModifiers().contains( Modifier.PRIVATE ) )
     {
-      throw new RouterProcessorException( "@" + type.getSimpleName() + " target must not be private", method );
+      throw new RouterProcessorException( "@" + toSimpleName( annotationName ) + " target must not be private",
+                                          method );
     }
   }
 
@@ -215,5 +223,96 @@ final class ProcessorUtil
 
       return true;
     }
+  }
+
+  @SuppressWarnings( "unchecked" )
+  @Nonnull
+  static List<AnnotationMirror> getRepeatingAnnotations( @Nonnull final Elements elements,
+                                                         @Nonnull final Element typeElement,
+                                                         @Nonnull final String containerClassName,
+                                                         @Nonnull final String annotationClassName )
+  {
+    final AnnotationValue annotationValue =
+      findAnnotationValue( elements, typeElement, containerClassName, "value" );
+    if ( null != annotationValue )
+    {
+      return ( (List<AnnotationValue>) annotationValue.getValue() ).stream().
+        map( v -> (AnnotationMirror) v.getValue() ).collect( Collectors.toList() );
+    }
+    else
+    {
+      final AnnotationMirror annotation = findAnnotationByType( typeElement, annotationClassName );
+      if ( null != annotation )
+      {
+        return Collections.singletonList( annotation );
+      }
+      else
+      {
+        return Collections.emptyList();
+      }
+    }
+  }
+
+  @SuppressWarnings( "SameParameterValue" )
+  @Nullable
+  private static AnnotationValue findAnnotationValue( @Nonnull final Elements elements,
+                                                      @Nonnull final Element typeElement,
+                                                      @Nonnull final String annotationClassName,
+                                                      @Nonnull final String parameterName )
+  {
+    final AnnotationMirror mirror = findAnnotationByType( typeElement, annotationClassName );
+    if ( null != mirror )
+    {
+      return findAnnotationValue( elements, mirror, parameterName );
+    }
+    else
+    {
+      return null;
+    }
+  }
+
+  @Nullable
+  private static AnnotationValue findAnnotationValue( @Nonnull final Elements elements,
+                                                      @Nonnull final AnnotationMirror annotation,
+                                                      @Nonnull final String parameterName )
+  {
+    final Map<? extends ExecutableElement, ? extends AnnotationValue> values =
+      elements.getElementValuesWithDefaults( annotation );
+    final ExecutableElement annotationKey = values.keySet().stream().
+      filter( k -> parameterName.equals( k.getSimpleName().toString() ) ).findFirst().orElse( null );
+    return values.get( annotationKey );
+  }
+
+  @Nonnull
+  static AnnotationValue getAnnotationValue( @Nonnull final Elements elements,
+                                             @Nonnull final AnnotationMirror annotation,
+                                             @Nonnull final String parameterName )
+  {
+    final AnnotationValue value = findAnnotationValue( elements, annotation, parameterName );
+    assert null != value;
+    return value;
+  }
+
+  @Nonnull
+  static AnnotationMirror getAnnotationByType( @Nonnull final Element typeElement,
+                                               @Nonnull final String annotationClassName )
+  {
+    AnnotationMirror mirror = findAnnotationByType( typeElement, annotationClassName );
+    assert null != mirror;
+    return mirror;
+  }
+
+  @Nullable
+  static AnnotationMirror findAnnotationByType( @Nonnull final Element typeElement,
+                                                @Nonnull final String annotationClassName )
+  {
+    return typeElement.getAnnotationMirrors().stream().
+      filter( a -> a.getAnnotationType().toString().equals( annotationClassName ) ).findFirst().orElse( null );
+  }
+
+  @Nonnull
+  private static String toSimpleName( @Nonnull final String annotationName )
+  {
+    return annotationName.replaceAll( ".*\\.", "" );
   }
 }