+++ processor/src/main/java/router/fu/processor/package-info.java
@@ -1,4 +1,4 @@
 /**
- * This package contains the processor for the {@link router.fu.annotations.Router} annotation.
+ * This package contains the processor for the router.fu.annotations.Router annotation.
  */
 package router.fu.processor;